Create shortcut command in node
When the project gets bigger and bigger, we may want to output or develop different versions or even run different tests. Each of these operations requires you to enter a series of commands in the command line with its own switches, which takes time. If you want to share the project, you must share these commands and switches with other programmers.
:alien:
npm run build